+++ eth_defi/confirmation.py
@@ -5,14 +5,21 @@ import logging
 import time
 from typing import Dict, List, Set, Union
 
+import rlp
 from eth_account.datastructures import SignedTransaction
 from hexbytes import HexBytes
 from web3 import Web3
 from web3.exceptions import TransactionNotFound
 
 from eth_defi.hotwallet import SignedTransactionWithNonce
+from eth_defi.tx import decode_signed_transaction
 
 logger = logging.getLogger(__name__)
+
+
+
+class BroadcastFailure(Exception):
+    """Could not broadcast a transaction for some reason."""
 
 
 class ConfirmationTimedOut(Exception):
@@ -127,6 +134,13 @@ def broadcast_transactions(
         Set to zero to return as soon as we see the first transaction receipt
         or when using insta-mining tester RPC.
     :return: List of tx hashes
+
+    :raise BroadcastFailure:
+        If the JSON-RPC node rejects the transaction.
+
+        - Anvil will reject some transactions immediately: if there is not enough gas money
+
+        - Ethereum Tester reject some transactions immediately on any error in automining mode
     """
     # Detect Ganache
     chain_id = web3.eth.chain_id
@@ -142,7 +156,14 @@ def broadcast_transactions(
     hashes = []
     for tx in txs:
         assert isinstance(tx, SignedTransaction) or isinstance(tx, SignedTransactionWithNonce), f"Got {tx}"
-        hash = web3.eth.send_raw_transaction(tx.rawTransaction)
+
+        try:
+            hash = web3.eth.send_raw_transaction(tx.rawTransaction)
+        except ValueError as e:
+            # Anvil/Ethereum tester immediately fail on the broadcast
+            # ValueError: {'code': -32003, 'message': 'Insufficient funds for gas * price + value'}
+            decoded_tx = decode_signed_transaction(tx.rawTransaction)
+            raise BroadcastFailure(f"Could not broadcast transaction: {decoded_tx}") from e
 
         assert hash