+++ eth_defi/hotwallet.py
@@ -16,7 +16,12 @@ logger = logging.getLogger(__name__)
 
 
 class SignedTransactionWithNonce(NamedTuple):
-    """Helper class to pass around the used nonce when signing txs from the wallet."""
+    """Helper class to pass around the used nonce when signing txs from the wallet.
+
+    - Emulate `SignedTransaction` from web3.py package
+
+    - Add some debugging helpers
+    """
 
     rawTransaction: HexBytes
     hash: HexBytes
@@ -24,6 +29,11 @@ class SignedTransactionWithNonce(NamedTuple):
     s: int
     v: int
     nonce: int
+
+    #: Undecoded transaction data as a dict.
+    #:
+    #: If broadcast fails, retain the source so we can debug the cause
+    source: Optional[dict] = None
 
     def __getitem__(self, index):
         return __getitem__(self, index)
@@ -87,6 +97,7 @@ class HotWallet:
             r=_signed.r,
             s=_signed.s,
             nonce=tx["nonce"],
+            source=tx,
         )
         return signed